Move MaterialLibraryState to engine from editor
Instanceserver is now importing it, can't be in the
editor package.
