+++ packages/editor/src/components/element/PropertiesPanelContainer.tsx
@@ -32,13 +32,13 @@ import { Entity, UndefinedEntity } from '@etherealengine/ecs/src/Entity'
 import { getMutableState } from '@etherealengine/hyperflux'
 import { UUIDComponent } from '@etherealengine/spatial/src/common/UUIDComponent'
 
+import { MaterialSelectionState } from '@etherealengine/engine/src/scene/materials/MaterialLibraryState'
 import { Popover } from '@mui/material'
 import { EntityNodeEditor } from '../../functions/ComponentEditors'
 import { EditorState } from '../../services/EditorServices'
 import { SelectionState } from '../../services/SelectionServices'
 import { PropertiesPanelButton } from '../inputs/Button'
 import MaterialEditor from '../materials/MaterialEditor'
-import { MaterialSelectionState } from '../materials/MaterialLibraryState'
 import { CoreNodeEditor } from '../properties/CoreNodeEditor'
 import ElementList from './ElementList'
 import { PopoverContext } from './PopoverContext'
